Burnable TBTC ERC20 token with permit
TBTC is now a burnable ERC20 token with EIP2612 permit functionality.
User can authorize a transfer of their token with a signature conforming
EIP712 standard instead of an on-chain transaction from their address.
Anyone can submit this signature on the user's behalf by calling the
permit function, as specified in EP2612, paying gas fees, and possibly
performing other actions in the same transaction.

By the time of implementing this contract, to my knowledgem there is no other
audited EIP2612 token implementation than Uniswap V2 on which this code is
based on. OpenZeppelin has quite recently added permit to their ERC20
but this code is in a draft state, was released in the most recent
v4.0.0, and to my understanding was not audited yet and may have
breaking changes in the future. On the contrary, Uniswap implementation
was audited and is even referenced from the EIP as an implementation
example.

This implementation mostly mirrors UniswapV2ERC20, and is identical to
the implementation we use in coverage pools. It adds some tweaks from
OpenZeppelin ERC20 implementation such as meaningful revert messages and
Approval event emitted on calls to transferFrom. This allows
applications to reconstruct the allowance for all accounts just by
listening to the event.

Unit tests were based on OpenZeppelin ERC20 unit tests with our own
tests implementation for EIP2612 and EIP712.

In the future, we will add more functionalities to the token, such as
approveAndCall and recovering tokens mistakenly sent to TBTC token
address.