Validate chain ID for every permit call
Calling UnderwriterToken.permit allows an asset holder to move tokens on behalf
of another user by using a signature. The lack of chain ID verification on every
signature could enable an attacker to withdraw rewards multiple times on behalf
of someone else if the chain was forked with a different chain ID.

The chain ID was read only at the deployment time and DOMAIN_SEPARATOR was
constructed based on that value. It could happen that the forked chain had
another chain ID, and given that we were not performing chain ID verification
for each permit call, some transactions could be replayed. The same issue
is in LP token v2 version this code is based on.

To address this problem, chain ID and DOMAIN_SEPARATOR constructed
for the given chain ID are cached and chain ID is validated for every permit
call. In case the cached value does not match the current one, a new
DOMAIN_SEPARATOR is constructed based on the current chain ID value.