Added approveAndCall to TBTC token
approveAndCall is not a part of ERC20 - it was specified in ERC827 and
slightly diverged for some implementations. This one is identical to
approveAndCall behavior of KEEP token.

approveAndCall executes receiveApproval function on the spender as
specified in IReceiveApproval interface. Approves spender to withdraw
from the caller multiple times, up to the value amount. If this function
is called again, it overwrites the current allowance with the provided
value. Reverts if the approval operation reverted or if receiveApproval
call on the spender reverted.