Separated update initiator and update finalizer roles
All governable parameters are now controlled by two roles: update
initiator and finalizer. There is a saparate initiator role for unmint
fee update and vending machine update, and finalizer is always the
contract owner. The initiator proposes the change by initating the
update and the finalizer (contract owner) may approve it by finalizing
the change after the governance delay passes.

The reason for this change were possible attack vectors on
single-address governance and what are the implications of them. Updated
vending machine address can run any code (it can even be an EOA), and
even if we decided not to transfer any TBTC v1 to the updated vending
machine, unmint fee update could make TBTC v2 > TBTC v1 not possible.