Use MisfundRecovery from @thesis/solidity-contracts
We extracted MisfundRecovery to @thesis/solidity-contracts so that we
can reuse it from keep-network/tbtc-v2 and
threshold-network/solidity-contracts.

The functionality stays as-is.

Also, worth mentioning, ERC20WithPermit from the updated commit hash
does not allow to transfering tokens to the token address. Most misfund
cases we've seen were about transfering KEEP to KEEP token address. This
can be prevented with one require added to ERC20WithPermit in the
updated commit hash.