Cache npm dependencies based on lock file in `solidity`
We were caching npm dependencies based on the `package-lock.json` file
located in the root directory. But that package is a small one and we
install it only in one job. It makes more sense to cache the
dependencies of the package located in the `solidity` directory, which
contains much more dependencies and is installed in each job of the
workflow.

Here's how long it takes to install dependencies in each job depending
on the caching methods:

Build-and-test:
Without cache: ~1m20s
With cache in `.`: ~1m37s
With cache in `./solidity`: ~42s

Format:
Without cache: ~2s+1m19s
With cache in `.`: ~1s+1m46s
With cache in `./solidity`: ~49s

Slither:
Without cache: ~1m45s
With cache in `.`: ~1m38s
With cache in `./solidity`: ~45s