Configure jobs deploying & publishing contracts
We're adding jobs which will deploy and publish `tbtc-v2` contracts.
Published artifacts will be used in the Continuous Integration proccess.
`NPM` workflow will build & publish contracts every time there is a
change in contracts or in the files describing package dependencies.
The `-dev` suffix will be used in the name of the published package.
`Solidity` workflow will deploy and publish contracts every time
worklfow is triggered by the `workflow_dispatch` event. Published
package name will point to name of the environment on which the
contracts were deployed.
Note: we're temporarily seitche of publishing to NPM registry using
`--dry-run` flag. The flag will be removed before merging of the changes
to the `main` branch.