RFC 5: Emptying and closing wallets
When a wallet reaches a maximum age,
falls below a minimum amount of held BTC,
falls below a minimum number of operators,
or fails to complete any operation within the allowed timeout,
it is to be emptied out with any remaining BTC transferred out.
The transaction emptying out a wallet is treated like a BTC redemption.