Move etherscan verification to a separate job.
We decided to move the etherscan verification to a seperate job,
executed after the job which migrates contracts. This way we can get rid
of the `continue-on-error` flag and adding annotations in case of
verification failure. With etherscan verification in separate job we can
allow the job to fail in case of problems with the veryfication - the
failure will not block anything and will be more visible than the
plain annotation would be.