Remove references to `wallet_max_btc`
This solidifies the design as having no cap on the amount of BTC any
particular wallet can contain. This simplifies the design, massivly
helps us scale, and helps us deal with whale-deposits. On the downside,
it exposes us to black-swan risk where if the adversary manages to steal
all of the BTC in a big wallet, the system can get absolutely blown out
since a wallet can be bigger than the whole covpool.