Bitcoin Bank Vault contract
Vault is a contract extended by applications built on top of the Bank. Vault is
locking balances under its own address by transferring the balances to the Vault
address and internally bookkeeping the amount of a locked balance per account.
This way, each application has a separate Bitcoin balance and problems in a
single application are not propagated to other applications or Bank depositors.