Do not pass the amount to Bank.increaseBalanceAndCall
Instead of passing the amount as a parameter, we can sum up all
`depositedAmounts` elements. This way, the application can be sure
that the amounts passed by the Bank align with the balance increase
for the application.

The gas cost is minimal: +447 gas units for 2 deposits and +1758
gas units for 5 deposits.