Validate array length in Bank.increaseBalanceAndCall
We do validate array lengths in Bank.increaseBalances so it makes
sense to do the same in Bank.increaseBalanceAndCall.

We do not validate if depositors array is non-empty in Bank but we
validate it in TBTCVault. There could be a scenario where some
vault wants to pass empty arrays because it does not care about
who deposited, e.g. some fundraising contract.