Refactor wallet BTC balance parameters
So far, there were two parameters responsible
for controlling the wallet creation and closure:
- `walletMinBtcBalance`
- `walletMaxBtcBalance`
The purpose of the latter parameter was
clear since it was used for closure exclusively
but the former was ambiguous as it was
referring both creation and closure. In order
to break that ambiguity, we refactor those
parameters to the following form:
- `walletCreationMinBtcBalance`
- `walletCreationMaxBtcBalance`
- `walletClosureMinBtcBalance`
Now, each parameter clearly refer to either
wallet creation or closing.