Reveal donation
This change exposes the `revealDonation` function
that allows to reveal an external BTC donation.
The donation logic lives in the `Donation.sol`
library. Once a donation is revealed to the
`Bridge`, a merging funds request is created
in order to notify the recipient wallet about
the need of merging the obtained UTXO with
the current main UTXO and update the BTC balance.
Merging funds mechanism is not exclusively
connected with donations and will be also used
to merge UTXOs received through the moving funds
process, hence the logic lives in a separate
`MergingFunds` library.