Extract `processPubKeyHashTxOutput` function
There is a common pattern in transactions that
locks funds on wallets' public key hashes. This
pattern is about taking the output, extracting
the 20-byte hash and making sure it is used in
the right context of a P2PKH or P2WPKH output.
Because of that, we extract this logic to
the `BitcoinTx` library and will reuse it in
all places that needs it.