Establish expected format for wallet ECDSA signing group heartbeat messages
The problematic scenario is when a group signs some message (say every nth block
hash) with their threshold ECDSA group public key, and then, malicious group
member uses the signed message as a sighash for a fraud challenge. Since that
sighash was never used on Bitcoin, the fraud challenge can not be defeated.

We address this problem by establishing a format for a heartbeat message making
sure that  message in this format can never become a sighash of Bitcoin
transaction.

We construct the heartbeat message, as a 40-byte message. The first 8 bytes are
0xffffffffffffffff. The last 32 bytes are for an arbitrary uint64, being
a signed heartbeat nonce (for example, the last Ethereum block hash).

This commit is just establishing the format of a heartbeat message. Integration
with Fraud library will be implemented separately.