Introduce moved funds sweep request state
It is important to know the final result of
the moved funds sweep request. It can be either
processed or timed out and this is important
for frauds handling that will be done in the
follow-up PR. Last but not least, the state
is more elegant to handle that two separate
timestamps indicating the final result.