Approve balance redemption
This changeset introduces the ability to perform
the transition from TBTC token to BTC within
a single transaction. To do so, a TBTC holder can
use the token's `approveAndCall` function to
request a redemption in the `Bridge` by passing
the required redemption parameters as `extraData`.

The entire flow looks as follows:
- `TBTC.approveAndCall` sets the allowance
for the `TBTCVault` and call their `receiveApproval`
function
- `TBTCVault.receiveApproval` burns the TBTC token
and transfers the redeemed amount from vault's
balance back to the redeemer's balance.
If `extraData` are set, there is an
additional call to `Bank.approveBalanceRedemption`
in order to approve the amount for redemption
- `Bank.approveBalanceRedemption` asserts the
balance is transferred, sets the allowance for
the `Bridge`, records the approving party and
notifies the `Bridge` about the approval
using `receiveBalanceRedemptionApproval`
- `Bridge.receiveBalanceRedemptionApproval`
makes sure the call is made by the `Bank`,
asserts the approver is a trusted vault or
the redeemer itself and parses the `extraData`.
Once done, it forms a new redemption request
using regular `requestRedemption` call. If the
redemption request hits the timeout, the redeemer
can get back their balance using the
`notifyRedemptionRequestTimeout` call