Extracted receiveBalanceApproval out of Vault to a separate interface
This way, also the Bridge will be able to implement this function. The idea
behind `Bank.approveBalanceAndCall` is that any smart contract can consume the
approved balances. In our system there are two options: the consumer is a `Vault`
or the consumer is the `Bridge`. The latter will let us implement more gas-efficient
redemption requests by placing redemption requests in a single transaction, the
same in which the balance is approved to the Bridge.