Set of updates to Bank contract documentation
Added `@params` to all public functions - this approach is aligned with Bridge
contract. Clearly listed all requirements in `@dev` section, and did a set of
grammar updates. Updated `increaseBalanceAndCall` documentation to do not
mention sweeping given it is implementation detail of the Bridge and not
something the Bank should not be interested in. Also. the documentation of that
function would get outdated when we would implement quickminting.