Bridge.receiveBalanceApproval
Previously, it was needed to execute two separate transactions to redeem Bank
balance to BTC: `Bank.approve` and `Bridge.requestRedemption`. The change
extracting `receiveBalanceApproval` out of `IVault` to a separate
`IReceiveApproval` interface allowed to integrate it with `Bridge` contract that
can now receive the approved balance by requesting redemption. This way, redeemer
can call `Bank.approveBalanceAndCall` with `Bridge` contract to request
redemption in a single transaction.

Next to Bitcoin-specific data passed as `bytes extraData` to
`approveBalanceAndCall /receiveBalanceApproval`, we also pass the redeemer
address that can be different from the `balanceOwner` address. A real-world use
of this functionality that will soon be implemented for `TBTCVault`
is when someone wants to go from vault balance straight to BTC. Then, the vault
is the balance owner but someone else is the redeemer who can claim the
balance back in case something goes wrong during the redemption.