Splitting BridgeGovernance into two files to limit the contract size
We have a lot of fields to put into the BridgeGovernance contract. It
creates a likely scenario where we hit the contract size limit before we
can add all the fields. To address this issue, a new lib is created
where we store these fields. It should also lift some function weight.