Added TBTCVault to deployment scripts
So far `TBTCVault` was not included in deployment scripts. To test
`TBTCVault<>Bridge` interactions we need to have it deployed.

The solution presented in this changeset is not transferring the ownership
of TBTC v1 token from `VendingMachine` to `TBTCVault`. This will have to be done
manually in selected unit tests. Transferring the ownership of TBTC v2 to
`TBTCVault` would make `VendingMachine` stop working. For now, we want to keep
it working for testnet KEEP Token Dashboard. Also, we need to implement a test
stressing `VendingMachine -> TBTCVault` token ownership transfer. Not having
this transfer done in deployment scripts should make implementing this test
easier.