One-transaction TBTC to BTC
Modified `TBTCVault.receiveApproval` function to allow unminting TBTC and
requesting Bank balance redemption in the Bridge in a single transaction via
`TBTC.approveAndCall`. Also, added `TBTCVault.redeem` function allowing to
unmint TBTC and request redemption in the `Bridge` on an already approved TBTC
balance.

During the work on this feature, I have been exploring different options,
including exposing `IBridge` interface that would be called from `Bank` or
`TBTCVault` to redeem balance. The current version won because it does not
tightly-couple vault implementations with the specific `Bridge` implementation.
For example, the second version of the `Bridge` may not even have a concept of
a wallet or a main wallet UTXO. Passing parameters as bytes is more flexible and
uses an existing `Bank` mechanism of `receiveBalanceApproval`.