Adding a lib storage with grouped data by type
Some of the helper methods like getRemaining_UpdateTime were had to be
removed because of the contract size. All these data is present in the
events though and can be fetched from there.

There are a lot of workarounds to resize the contract. Implementing in a
'normal' way would increase the size to around 32KB which is way above
the limit.