Introduced resolveRedeemingWallet, removed main UTXO validation from OutboundTx
Main UTXO validation was a part of `OutboundTx.processWalletOutboundTxInput`.
This change removes the main UTXO validation from there.
For `submitRedemptionProof`, the validation happens in the newly introduced
`resolveRedeemingWallet`. For `submitMovingFundsProof`, the validation happens
inside `submitMovingFundsProof`.

First, this change allows saving gas on `submitMovingFundsProof` and
`submitRedemptionProof` given we do not fetch wallet pointer twice from the
mapping. Before/after:

| Bridge · submitMovingFundsProof · 200904 · 356450 · 304175 │
| Bridge · submitRedemptionProof · 130155 · 208356 · 176074 │

| Bridge · submitMovingFundsProof · 200900 · 356446 · 304171 │
| Bridge · submitRedemptionProof · 130103 · 208304 · 176025 │

Second, it allows to align wallet state validation rules.

If the function is not performing a state transition, the wallet state should be
validated inside that function. For most functions (for example
`Deposit.revealDeposit`, `Redemption.requestRedemption`) this will be the very
first line of the logic.

There are three exceptions to this rule: `submitDepositSweepProof`,
`submitRedemptionProof`, `submitMovedFundsSweepProof`.
For `DepositSweep.submitDepositSweepProof` validating wallet state inside
`submitDepositSweepProof` leads to stack too deep problems. Extracting it to
a separate function helps. `Redemption.submitRedemptionProof` and
`submitMovedFundsSweepProof` are aligned and they also perform the validation
when resolving the wallet. To make it easier for future-us and external
contributors, all three functions have now a comment informing where the state
is validated.