Changed the order of functions in Wallets library
`getWalletBtcBalance` utility view function is the last one in the file given it
has nothing to do with wallet state changes.

`moveFunds`, `beginWalletClosing`, `finalizeWalletClosing`, and `terminateWallet`
are called only from `Wallets` library and they are at the end of the file, just
before `getWalletBtcBalance`.

The rest of the functions was ordered based on the state transition:
- wallet creation
- timed out redemption
- heartbeat failed
- closing wallet
- moving funds
- fraud

NO LOGIC CHANGES, JUST UPDATING THE ORDER OF FUNCTIONS.