Assert emitted events, reset relay mock properly, capture initial balance once
Set of updates to MaintainerProxy unit tests.

First, we are now asserting not only whether the transaction did not revert but
whether it did emit correct event. This is much stronger guarantee that covers
a try-catch case as well.

Second, we were not resetting relay mock properly in "run scenario" functions.
We are doing it now.

Finally, we can capture initial authorized maintainer balance only one time
given all tests operate on clean state and are independent from each other.