Allow only trusted SPV maintainers to submit SPV proofs
The SPV proof does not check whether the transaction is a part of the Bitcoin
mainnet, it only checks whether the transaction has been mined performing the
required amount of work as on Bitcoin mainnet. The system then assumes that if
this is the case, the transaction is a part of the Bitcoin mainnet. In other
words, for 6 confirmations required, anyone who can produce 6 properly chained
blocks with the current or previous Bitcoin network difficulty, can prove
Bitcoin transactions to the system.

The possibility of submitting SPV proofs is now limited to trusted SPV maintainers.
The system expects transaction confirmations with the required work accumulated,
so trusted SPV maintainers can not prove the transaction without providing the
required Bitcoin proof of work.

Trusted maintainers address the issue of an economic game between tBTC and
Bitcoin mainnet where large Bitcoin mining pools can decide to use their hash
power to mine fake Bitcoin blocks to prove them in tBTC instead of receiving
Bitcoin miner rewards.

Running a full relay on L1 or L2 does not offer the proof of inclusion and is
also vulnerable to similar attacks as long as there is an incentive for mining
pools to mine fake blocks to use them against TBTC instead of using their hash
power to earn Bitcoin miner rewards.

For the launch, we will limit the possibility of submitting SPV proofs to
trusted maintainers. Note that, just like in v1, the system expects 6
confirmations with the required work accumulated, so trusted maintainers can not
prove the transaction without providing the required Bitcoin proof of work.

For future releases we will research alternatives to SPV proofs.

The work in this commit is incomplete. There is one SPV maintainer address
defined for the deployment and both this address and `MaintainerProxy` contract
can submit SPV proofs. In the next commit, `MaintainerProxy` will be refactored
to recognize separate category of maintainers - those eligible to submit SPV
proofs.