Separate a concept of a wallet and SPV maintainer in MaintainerProxy
In the previous commit, we updated the Bridge contract to let only SPV
maintainers submit SPV proofs. Here, `MaintainerProxy` is modified to separate
the role of SPV maintainer and wallet maintainer. There is nothing stopping the
governance to have the same list of SPV maintainers and SPV maintainers, but
this approach is more flexible.

There are two options now:
- SPV maintainer authorized in `Bridge` contract calling functions directly and
  paying gas: scenario similar to a v1 relay maintainer and ETH expense refunds
  from the DAO
- SPV maintainer authorized in `MaintainerProxy`, calling functions through the
  proxy and being reimbursed from the `ReimbursementPool`.