Address bitcoin timestamp irregularities
Bitcoin timestamps have the following constraints:
- they must be greater than the median of the previous 11 timestamps
- they must be less than the network-adjusted time +2h

Because of this, timestamps are not guaranteed to be strictly increasing
between blocks. This causes challenges in identifying the correct epoch
as the prior implementation relied on timestamps being ordered.

To address this, we identify the epoch we expect the block to belong in
based on the block's timestamp compared to the epochs' timestamps. If
there is an irregularity and the block does not fall within the expected
epoch, we can identify this by comparing the block's target to the epoch
target. In the event of a mismatch we check the previous and next epoch
for a matching target. If the target matches one of these, we consider
the epoch identified. If it matches neither, we consider the block's
target invalid.