Add sparse relay
The sparse relay records every sixth bitcoin block. To validate
individual transactions, we need a chain of at least six block headers.
The relay checks the internal validity of the chain, and finds if any of
the blocks have been recorded in the relay. If a match is found, the
relay checks that it is a part of the current longest chain, instead of
an orphaned one, and that a sufficient number of confirmations have been
mined on top of it.