Remvoing development env from `deployments` option
Development env is used for local development on network other than
hardhat. Currently it is Geth. Having development env in deployments
field prevents from proper deployment of tbtc contracts on Geth. The
transaction won't be completed and the contract won't be deployed on
Geth. Removing it, so we can properly deploy tbtc contracts.