Extracted BridgeFraudStub contract
“Evil is evil, Stregobor. Lesser, greater, middling, it's all the same.
Proportions are negotiated, boundaries blurred. I'm not a pious hermit,
I haven't done only good in my life. But if I'm to choose between one evil and
another, then I prefer not to choose at all.”
Geralt from Rivia

Before we can perform fixes on the target wallets commitment reimbursement, we
need to reduce the size of `BridgeStub` contract. Extracting three functions
used only for fraud tests to a separate contract is the option with the least
changes to the post-audited code from all the options I've been exercising.

Thanks to this change, `BridgeStub` reduces its size from 24.570 to 23.649
(by 0.921). `BridgeFraudStub` is still at capacity, but we can bypass this
problem by reducing the number of optimizer runs on `BridgeFraudStub`. I chose
not to reduce the number of optimizer runs on `BridgeStub` because it is the
contract deployed for the hardhat network used in tests. We need to use the same
settings as for mainnet for gas reporter and for unit tests, as much as possible.
If we do not do it just for `BridgeFraudStub`, this is the lesser evil that I
preferred to choose here.