Add a mock fn that register the ECDSA wallet
To create a deposit or request redemption the ECDSA wallet is required.
To not involve the off-chain clients here we add a mock function that
registers the fake ECDSA wallet in the `Bridge` contract. Anyone can
execute this function. In this mocked function we also pass the main
utxo params because it is required for redemption flow.