Add a mock function that sweeps a deposit
The off-chain clients are responsible for sweeping deposit. We need to
create mocked version of this function and we will trigger it manually-
we will mock the client work. To simplify the deposit sweep flow this
mocked function will take only one deposit. In taht fact we can mock the
off-chain clients work and sweep a deposit.