Does not allow updating an existing allowance to a non-zero value
Does not allow updating an existing allowance to a value that is non-zero to
avoid someone using both the old and the new allowance by unfortunate
transaction ordering. To update an allowance to a non-zero value one should
update to zero first or use `increaseBalanceAllowance` or
`decreaseBalanceAllowance` for an atomic update. We have been warning about this
case in `approveBalance` documentation, but we were not enforcing this behavior.
Now we have a `require` making sure the rule is obeyed.