Update solidity config for ecdsa
Update number of optimizer runs to reduce the size of the
`WalletRegistry` contract. Lower values of runs will optimize more for
initial deployment cost(contract size) but the execution costs will be
higher. We don't have to worry about the execution cost of
`WalletRegistry` contract because changes from `dapp-development` branch
are only for testnet networks.