Improve blinding factor explanation
During the course of the audit, auditors stated that if an attacker can operate
a dedicated hardware device at 100 mh/s, he can explore a 64-bit space in
~200,000 days and guess someone's blinding factor.

It is worth noting the `blindingFactor` (64 bits) is not the only parameter that
changes for every deposit: `refundLocktime` (32 bits) changes as a timestamp
for every deposit, and walletPubKeyHash (160 bits) rotates according to the
rules of an active wallet rotation.

All the mentioned values are a part of DepositRevealInfo so the space to explore
is not 64-bit but 64+32=96 bits assuming the active wallet is known. The active
wallet is rotating so there is a limited time to explore a 96-bit space and
there is absolutely no financial incentive for anyone to do that as they gain
nothing and make no harm to the depositor. They will pay gas for something the
depositor intended to do anyway. The analogy is front-running someone to call
a smart contract’s function that has the same effect no matter the caller. The
blinding factor is used to distinguish deposits from the same depositor.

We improved the documentation to make it clear the blinding factor is not
a security-critical parameter.