Don't run unit/integration tests for `dapp-development` PR
We don't want to run unit/integration tests on contracts modified to help with
testing of the dashboard (stored on a `dapp-development` branch). We already
were turning off the execution of such tests in case when `github.ref` is
`refs/heads/dapp-development` (i.e. when manual dispatch is executed on the
`dapp-development` branch). But the tests were still being executed as checks on
the PR for `dapp-development` branch, which is not a desired behaviour.
This is because PRs have different syntax of `github.ref`:
`refs/pull/<pr_number>/merge`. To turn off the execution of tests for PR on
`dapp-development` branch, we use condition involving `github.head_ref`, which
is a variable available only for `pull_request` or `pull_request_target` events.