Update LightRelay deployment script
We are going to use `LightRelay` contract as `IRelay` implementation.
The contract requires a relay maintainer to be working on the given
environment to provide the Bitcoin chain data.

We're not ready yet with the maintainer, so we can't use the
`LightRelay` on goerli. Hence, we kepp on using `TestRelay` which is a
dummy implementation of the `IRelay` interface.

Once the maintainer is ready we will switch goerli to use `LightRelay`.

Another thing is that we want just one instance of the `LightRelay`
contract to be deployed on the given environment. We don't want to
redeploy the `LightRelay` on goerli to update the maintainer and loose
all previousl submitted bitcoin chain data. To cover that we will expect
the LightRelay artifact will be committed to the repository after it's
deployed. This way we will reuse the existing instance of the contract.