Refactor assembleDepositScript function
There are two new types:
- DepositScriptParametersWithRefundPKHAddress
- DepositScriptParametersWithPublickKey

DepositScriptParameters type can be either first or second, so that means that
it can either have publick hey OR PKHAddress. Never both, and never none of
them.

The code that checks which parameter was used and validates the deposit script
params was extracted to separated function - buildDepositScriptHash - which is
one level up of assembleDepositScript. If the public key was used then we hash
it into PKHAddress which is then used in assembleDepositScript function.