RFC 7: Sequential signing instead of parallel signing
We experimented with parallel signing using GG18 protocol and it turned out to
be really unstable. We were not able to have a successful signing execution for
a batch of 10 messages. The problem is about very long periods of computations
with a lot of CPU spikes that can starve all other processes, for example,
networking. This makes the clients unresponsive and affects the execution of
all protocols which is unacceptable for production use. Moreover, the code
changes required by this approach have a negative impact on the overall code
quality and significantly increase the complexity level.