RFC 1: Removed "skip the sweeping line" section
We decided not to implement such functionality in the client but rather
provide a market-based solution where holders of TBTC can offer a
quickminting capability for some fee. For example, the depositor will
mint to the holder's escrow, and the holder will provide them with
holder's TBTC with some conditions applied.