RFC-7: Adjusted the sweeping schedule
The Coordinator may propose a sweep at any time if either `sweep_period` seconds
passed since the proof of the last sweep was accepted by the Bridge or
`sweep_max_deposits` waits in the queue for sweeping. The time it takes to sign
a sweep transaction input and Ethereum block gas limits put a cap on how many
deposits can be swept at once. Given that in normal circumstances, the active
wallet is not redeeming or moving funds, the active wallet can focus on sweeping
and optimize its work schedule for this operation, depending on the current
situation in the Bridge.