Run Solidity workflow also after changes to workflow files
We're changing workflow config so that the workflow would be triggered also on
changes to its config. This way we should be able to easily see if changes to
the workflow don't break anything (note that this does not apply to jobs that
are triggered only by the `workflow_dispatch` event - as this job publishes
important packages, we need more thoughtful approach to testing the changes
there).