Adjusted unit tests after TBTCVault optimistic minting change
TBTCVault inherits from TBTCOptimisticMinting now and it requires to pass
a reference to Bridge to the constructor.

In case of Bank.test.ts, the change was simple: pass signer-based Bridge.
No function is called on Bridge from Bank.test.ts so this simple solution
is just fine.

In case of TBTCVault.test.ts, smock-based mock of Bridge is used. Although
this would work with signer-based Bridge as well now, we will soon have more
tests for optimistic minting added and this require calls to Bridge.deposits()
that would fail for signer-based Bridge mock.