Do not allow to request optimistic mint twice
`pendingOptimisticMints` is intentionally cleared in `cancelOptimisticMint` to
cover the case of an errant Guardian who could cancel optimistic minting for
a valid revealed deposit. The idea was to leave the responsibility of canceling
optimistic mints if they were already finalized in the past to Guardians, but
given we can fit two timestamps (and more) into a single EVM word, we can have
this validation on-chain and make the Guardian implementation way lighter.