Accept zero Bridge deposit treasury fee
We have a condition in `Deposit.sol` in the `revealDeposit` function allowing
to properly handle the case when the Bridge's deposit treasury fee is zero:

```
deposit.treasuryFee = self.depositTreasuryFeeDivisor > 0
    ? fundingOutputAmount / self.depositTreasuryFeeDivisor
    : 0;
```

At the same time, we were not allowing for a zero `depositTreasuryFeeDivisor`
parameter in the validation of update parameter value functions.

This has been fixed and zero `depositTreasuryFeeDivisor` is allowed now. I have
also added unit tests ensuring revealing and sweeping deposit with a zero
treasury fee works fine.