Optimistic minting fee
The Optimistic Minting mechanism may additionally cut a fee from the amount that
is left after deducting the Bridge deposit treasury fee. Think of this fee as an
extra payment for faster processing of deposits. One does not need to use the
Optimistic Minting mechanism and they may wait for the Bridge to sweep their
deposit if they do not want to pay the Optimistic Minting fee.

The Optimistic Minting fee may be used by the Treasury for the donation to
"refill" the Bridge after the imbalance created after the finalized optimistic
minting requests with no Bitcoin sweep transaction miner fee taken into account.