Fixed unit tests for optimistic minting fees
Some of the tests to pass required changes from another PR fixing Bridge
update parameter functions to accept zero deposit and redemption fees.
The change was merge to main and main was reverse-merged to this branch
in the previous commit. Having the code fixed on the Bridge side, I
realized the tests have other reasons to fail, all fixed in this commit.