Get WalletRegistry artifact from @keep-network/ecdsa package
The @keep-network/ecdsa package is the source of the WalletRegistry
contract artifact that should be used in the code.

The artifact will not be available in the @keep-network/tbtc-v2 package.
It worked for `development` package as this is a specific of the
development package to contain artifacts for contracts from all the
modules.