Update downstream dependendencies
We're updating version of the downstream `@keep-network/random-beacon` and
`@keep-network/ecdsa` dependencies. This is because we want to always use this
version (or higher), as that version was built on Node.js v18 and v18 is what
we want to be using for building the `tbtc-v2` from now on as well.