Change SemVer range for `@tenderly/hardhat-tenderly` dependency
We can't use `@tenderly/hardhat-tenderly` in versions `1.4.x` and higher on
Node.js `v14`, as those `1.4.x` versions introduce dependency to `tslog`
package, which is incompatible with `v14` of Node.js (works on `v16` or higher).
There are also some problems with deployment scripts (possibly fixable) on
`@tenderly/hardhat-tenderly` in versions `1.2.x`, so we're limiting the allowed
versions to `<1.2.0`.