Set initial deposit reveal ahead period
For the initial launch with optimistic minting, we want a long deposit locktime
equal to 9 months. This locktime is set by the client code, but we also
need to enforce it on the contract side using the `depositRevealAheadPeriod`
parameter. We set that param to 8 months and 2 weeks. That actually gives
2 weeks to reveal a deposit with a locktime of 9 months.