Generate types for contracts with TypeChain
Use `typechain` to generate types for contracts artifacts.
We generate typings only for specific contracts that we use in this
library.

As a workaround for a bug https://github.com/dethcrypto/TypeChain/issues/772
we run `typechain` command separately for each ecdsa and tbtc-v2
projects and do `rm ./typechain/index.ts` at the end. This workaround
should be removed once a fix is ready.