Deploy LightRelay for all networks but goerli and system_tests
The tBTC Go client can be built for four networks/environments:
- `make all environment=local`: Build with contract packages deployed locally.
- `make all environment=development`: Build with contract packages published to
  the NPM registry and tagged `@development`.
- `make all environment=goerli`:  Build with contract packages published to the
  NPM registry and tagged `@goerli`.
- `make all environment=mainnet`: Build with contract packages published to the
  NPM registry and tagged `@mainnet`.

No matter the environment, we have an assumption in the Go code: the Ethereum
implementation for the relay chain expects the contract to expose some functions
available on the `LightRelay` contract.

It means that for all published NPM packages, tagged with `@mainnet`, `@goerli`,
or `@development`, we need to have the `LightRelay` contract artifact.

Goerli is a special case. Since on testnet we work against Bitcoin testnet where
the difficulty often drops to 1, a special variation of `LightRelay` contract
must be deployed there: `GoerliLightRelay`. That contract extends the
`LightRelay` so tBTC Go client generates the required bindings properly and the
previous and current epoch difficulty is always 1.

System tests are another special case. System tests use functions in the stub
relay contract to `setCurrentEpochDifficultyFromHeaders` and
`setPrevEpochDifficultyFromHeaders`. System tests - in the current form - do not
require tBTC Go client to run. For this reason, we deploy there
a `SystemTestRelay` for them. I have renamed the `TestRelay` contract to
`SystemTestRelay` to make it supper-explicit it is used only for system tests
and named the network that is supposed to be used for system tests as
`system_tests` network.

Unit tests work just fine with `LightRelay` deployed because they use Smock fake.