Add script for Bridge proxy upgrades
In case we need to upgrade Bridge contract here is the script.

The script expectes `Bridge.json` deployment artifact to be available in
`deployments/<network>` directory with the proxy contract details.
It also expected the new implementation contract to be named `Bridge`.

The script will use the `Bridge` contract, deploy and use as
implementation for the `Bridge` proxy. At the end the script will output
a deployment artifact that will overwrite the previous `Bridge`
deployment artifact.

When running it remove `func.skip` line and run the script with:
yarn deploy --tags UpgradeBridge --network <NETWORK>