Set `GOVERNANCE_DELAY` in `VendingMachine` to `0`
To QA the optimistic minting in T dapp we need to transfer the ownership
of the `TBTC` token contract. The `TBTCVault` should be an owner of the
`TBTC` token contract. The only way to do this is via `VendingMachine`
contract because is currently set as an owner of `TBTC` token contract
in deployment scripts.