Reverse function transaction hash
The TransactionHash type expects to hold a hash in the same order as
block explorers (RPC order). But the Bridge contract expects a hash in
the format that is used by bitcoin natively (internal byte order). The
event emitts the hash in the internal byte order, so we need to reverse
it.