Use Hex to handle hexadecimal values
The chain interface expects hexadecimal values to be stores as
un-prefixed strings. The Ethereum handles such values with `0x` prefix
by default. We need to convert the values to conform with the interface.

TODO:
To make dveloment easier and prevent from unexpected runtime errors we
need to handle the properties with Hex type.