Convert Bitcoin precision to Ethereum precision in TBTC Vault
This is an in-progress work and unit tests need to be fixed.

When minting TBTC we should multiply the Bank balance expressed in satoshis by
`SATOSHI_MULTIPLIER = 10**10` to convert from Bitcoin precision to Ethereum
precision. `mint`, `unmint`, `unmintAndRedeem` and `receiveApproval` functions
of `TBTCVault` operate in Ethereum world so they accept 1e18 precision numbers.
`receiveBalanceApproval` and `receiveBalanceIncrease` are called by the Bank
which operates in satoshis so they accept Bitcoin precision. An important
requirement when converting between precisions is to strip the amount so that
the change can not be used for printing tokens.

`strip1e18Amount` view functions allows to check the stripped amount, remainder,
and the Bank balance.