Introduce temporary retry mechanism to system tests Electrum calls
The testnet Electrum instances are often unreliable and cause failures
of system test scenarios. Here we aim to mitigate that problem by wrapping
all Electrum calls done by system tests with a temporary retry mechanism.
Eventually, this mechanism should be replaced with a full-fledged retry
support built-in in the `ElectrumClient` at the `tbtc-v2.ts` level.