Rerun System Tests on failure
We've seen that from time to time System Tests workflow fails due to problems
with connectivity to Electrum or due to some other randomly occurring issues
(like the problems with installing dependencies). We now add a job to the System
Tests workflow which is started only if one of the frevious jobs failed. The job
will trigger a Rerun workflow, which will then trigger the System Tests again.
The reason why we need to rerun via a separate workflow is because reruns cannot
be called on the workflows/jobs which are still running.