Block checkpoint mechanism
Here we add the logic that determines the block window for the monitoring
run based on the persisted block checkpoint. We use GCS bucket as persistence
medium. Checkpoint is updated only in case of a successful call so if
something goes wrong, the next try has the failed block window in scope.
That means no events are lost but duplicates are possible.