Limit number of workflow attempts to 3
We've been implementing workflow retries as a solution to some non-deterministic
failures of the 'System Tests' workflow that are not indicative of the problems
with the tested functionalities. The solution that we have built so far could
lead to constant rerty loop in case of consisting failures.
Now we're adding a retry limit - we will not automatically execute more than 3
attempts of the workflow.