L2 generic token: add and remove minter
This is the first step for the implementation of a generic L2/sidechain token
contract.

The L2/sidechain token will be upgradeable, owned by Threshold Council, will
delegate the minting authority to multiple parties, and have a pause
functionality for mints and burns. Individual L2/sidechain token implementations
will inherit from this contract setting only the token name and symbol.

So far, adding and removing minters was implemented. More to come!