Updated OpenZeppelin dependency to version 4.8.1
4.8.1 is the most recent version. In the `main` version of OpenZeppelin,
`draft-ERC20PermitUpgradeable` is no longer a draft but this change is
not yet present in 4.8.1 version of OpenZeppelin. Once a new version
gets released we should update to drop the `draft-` prefix from the
import in L2TBTC (see the very next commit). For now, I am updating to
the most recent version to be closer to 4.8.2 or 4.9.0 we will be
updating to.