L2TBTC token tests for ERC20 functionalities
Only the functions defined in L2TBTC are fully covered with tests. L2TBTC
contract inherits from OpenZeppelin contracts and we do not want to test the
framework. The basic tests for functions defined in the OpenZeppelin contracts
ensure all expected OpenZeppelin extensions are inherited in L2TBTC contract and
that they are properly initialized.